backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML
openai==1.3.8
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# Vector Database
pinecone-client==2.2.4

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# CORS
fastapi-cors==0.0.6 